chore: Delete pre-redesign views directory and move emails
