+++ .eslintrc.json
@@ -126,7 +126,6 @@
         "src/app.js",
         "src/utils/redisMock.js",
         "src/routes/**/*.js",
-        "src/views/**/*.js",
         "src/middleware/**/*.js",
         "src/controllers/**/*.js",
         "src/app/functions/server/breachResolution.ts"
