+++ src/app/api/utils/auth.ts
@@ -17,8 +17,8 @@ import { getBreaches } from "../../functions/server/getBreaches";
 import { getBreachesForEmail } from "../../../utils/hibp.js";
 import { getSha1 } from "../../../utils/fxa.js";
 import { getEmailCtaHref, initEmail, sendEmail } from "../../../utils/email.js";
-import { getTemplate } from "../../../views/emails/email2022.js";
-import { signupReportEmailPartial } from "../../../views/emails/emailSignupReport.js";
+import { getTemplate } from "../../../emails/email2022.js";
+import { signupReportEmailPartial } from "../../../emails/emailSignupReport.js";
 import { getL10n } from "../../functions/server/l10n";
 import { OAuthConfig } from "next-auth/providers/oauth.js";
 import { SerializedSubscriber } from "../../../next-auth.js";
