+++ src/emails/emailBreachCard.js
@@ -2,9 +2,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { getLocale, getStringLookup } from '../../utils/fluent.js'
-import { formatDate } from '../../utils/formatDate.js'
-import { getBreachLogo } from '../../utils/breachLogo.js'
+import { getLocale, getStringLookup } from '../utils/fluent.js'
+import { formatDate } from '../utils/formatDate.js'
+import { getBreachLogo } from '../utils/breachLogo.js'
 
 const breachAlertTableStyle = `
   margin: auto
