+++ src/emails/emailMonthlyUnresolved.js
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { getMessage, getStringLookup } from '../../utils/fluent.js'
+import { getMessage, getStringLookup } from '../utils/fluent.js'
 
 const emailStyle = `
   color: black;
