+++ src/emails/emailVerify.js
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { getStringLookup } from '../../utils/fluent.js'
+import { getStringLookup } from '../utils/fluent.js'
 
 const containerStyle = `
   background: #f9f9fa;
