+++ src/scripts/emailBreachAlerts.js
@@ -22,8 +22,8 @@ import {
   addEmailNotification,
   markEmailAsNotified,
 } from "../db/tables/email_notifications.js";
-import { getTemplate } from "../views/emails/email2022.js";
-import { breachAlertEmailPartial } from "../views/emails/emailBreachAlert.js";
+import { getTemplate } from "../emails/email2022.js";
+import { breachAlertEmailPartial } from "../emails/emailBreachAlert.js";
 import {
   initEmail,
   EmailTemplateType,
