--- src/views/partials/admin.js
@@ -1,17 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-/**
- * @param {unknown} _data
- * @returns string
- */
-export const admin = _data => `
-<section>
-  <h1>Admin</h1>
-
-  <ul>
-    <li><a href='/admin/emails'>Email preview</a></li>
-  </ul>
-</section>
-`
